refactored About Page file structure
