+++ src/components/about/WorkAccordian/WorkAccordian.css
